Address Copilot review comments: restore 4.4 endpoint behavior
Copilot comment:

The old-only path is not preserving prior behavior here. `SniEndPoint` is not a `DefaultEndPoint` (the former extractor branches were mutually exclusive), so this getter previously returned `null` for SNI; this block newly adds proxy-valued `network.peer.*` attributes in the frozen old conventions. Please retain the former null result for old-only SNI; only `server.*` previously used the reflected proxy address.
```
    if (!emitStableDatabaseSemconv()) {
      // The old database semantic conventions are frozen, so preserve the proxy as the peer.
      return getProxyAddress((SniEndPoint) endPoint);
```

Copilot comment:

This changes non-SNI custom endpoints, although the PR states that direct endpoints keep their exact behavior. The previous 4.4 extractor handled only `DefaultEndPoint` and `SniEndPoint`; arbitrary endpoints emitted no `server.*` attributes. Calling an unknown implementation's `resolve()` here both adds telemetry and can introduce implementation-specific resolution side effects unrelated to the SNI fix. Please leave unknown endpoint types unchanged.
```
    } else {
      SocketAddress address = endPoint.resolve();
      if (address instanceof InetSocketAddress) {
        attributes.put(SERVER_ADDRESS, ((InetSocketAddress) address).getHostString());
        attributes.put(SERVER_PORT, ((InetSocketAddress) address).getPort());
```

Copilot comment:

This broadens the 4.4 peer-address path from `DefaultEndPoint` to every non-SNI implementation. Custom endpoints previously returned no `network.peer.*`; now each span invokes arbitrary `resolve()` logic and emits new attributes, contradicting the PR's guarantee that non-proxied behavior is unchanged. Keep the existing `DefaultEndPoint` behavior and return null for unknown endpoint types.

This issue also appears on line 74 of the same file.
```
    if (!(endPoint instanceof SniEndPoint)) {
      SocketAddress address = endPoint.resolve();
      return address instanceof InetSocketAddress ? (InetSocketAddress) address : null;
```

Analysis: The prior batch incorrectly applied the Cassandra 4.0 compatibility model to the 4.4 module. The 4.4 baseline emitted network peer attributes only for DefaultEndPoint and emitted server attributes only for DefaultEndPoint and SniEndPoint. It also left SNI network peer attributes unset in every semantic-convention mode. This change restores those boundaries and updates the pull request description to state each module's exact old-only and custom-endpoint behavior.

Upsides: Cassandra 4.4 no longer adds attributes or invokes resolve() for unknown endpoint implementations. SNI network peer attributes remain unchanged in old-only mode, while the stable server-address fix remains intact.

Downsides: Cassandra 4.4 continues to omit network peer attributes under old-only SNI and to omit attributes for unsupported custom endpoints, matching its prior behavior.
