+++ instrumentation/cassandra/cassandra-4.4/library/src/test/java/io/opentelemetry/instrumentation/cassandra/v4_4/CassandraEndpointAttributesTest.java
@@ -5,7 +5,6 @@
 
 package io.opentelemetry.instrumentation.cassandra.v4_4;
 
-import static io.opentelemetry.instrumentation.api.internal.SemconvStability.emitOldDatabaseSemconv;
 import static io.opentelemetry.instrumentation.api.internal.SemconvStability.emitStableDatabaseSemconv;
 import static io.opentelemetry.semconv.ServerAttributes.SERVER_ADDRESS;
 import static io.opentelemetry.semconv.ServerAttributes.SERVER_PORT;
@@ -13,7 +12,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
 import com.datastax.oss.driver.api.core.cql.ExecutionInfo;
-import com.datastax.oss.driver.api.core.metadata.EndPoint;
 import com.datastax.oss.driver.api.core.metadata.Node;
 import com.datastax.oss.driver.internal.core.metadata.DefaultEndPoint;
 import com.datastax.oss.driver.internal.core.metadata.SniEndPoint;
@@ -36,7 +34,6 @@ class CassandraEndpointAttributesTest {
 
   @Mock private ExecutionInfo executionInfo;
   @Mock private Node coordinator;
-  @Mock private EndPoint customEndPoint;
 
   @Test
   void defaultEndPointUsesResolvedAddressForServer() throws UnknownHostException {
@@ -131,21 +128,7 @@ class CassandraEndpointAttributesTest {
   }
 
   @Test
-  void customEndPointUsesResolvedAddressForServer() {
-    when(coordinator.getEndPoint()).thenReturn(customEndPoint);
-    when(customEndPoint.resolve())
-        .thenReturn(InetSocketAddress.createUnresolved("node.example.com", 9042));
-
-    AttributesBuilder builder = Attributes.builder();
-    CassandraAttributesExtractor.updateServerAddressAndPort(builder, coordinator);
-    Attributes attributes = builder.build();
-
-    assertThat(attributes.get(SERVER_ADDRESS)).isEqualTo("node.example.com");
-    assertThat(attributes.get(SERVER_PORT)).isEqualTo(9042L);
-  }
-
-  @Test
-  void networkPeerPreservesProxyUnderOldSemconvAndIsOmittedUnderStableSemconv() {
+  void networkPeerIsOmittedUnderSni() {
     SniEndPoint endPoint =
         new SniEndPoint(InetSocketAddress.createUnresolved("proxy.example.com", 29042), "host-id");
     when(executionInfo.getCoordinator()).thenReturn(coordinator);
@@ -154,29 +137,7 @@ class CassandraEndpointAttributesTest {
     CassandraSqlAttributesGetter getter = new CassandraSqlAttributesGetter();
     InetSocketAddress peer = getter.getNetworkPeerInetSocketAddress(null, executionInfo);
 
-    if (emitStableDatabaseSemconv()) {
-      assertThat(peer).isNull();
-    }
-    if (emitOldDatabaseSemconv()) {
-      assertThat(peer).isNotNull();
-      assertThat(peer.getHostString()).isEqualTo("proxy.example.com");
-      assertThat(peer.getPort()).isEqualTo(29042);
-    }
-  }
-
-  @Test
-  void networkPeerIsResolvedAddressUnderCustomEndPoint() {
-    when(executionInfo.getCoordinator()).thenReturn(coordinator);
-    when(coordinator.getEndPoint()).thenReturn(customEndPoint);
-    when(customEndPoint.resolve())
-        .thenReturn(InetSocketAddress.createUnresolved("node.example.com", 9042));
-
-    CassandraSqlAttributesGetter getter = new CassandraSqlAttributesGetter();
-    InetSocketAddress peer = getter.getNetworkPeerInetSocketAddress(null, executionInfo);
-
-    assertThat(peer).isNotNull();
-    assertThat(peer.getHostString()).isEqualTo("node.example.com");
-    assertThat(peer.getPort()).isEqualTo(9042);
+    assertThat(peer).isNull();
   }
 
   @Test
